Simplify Storage interface
Rather than treating block data and state data as fundamentally
different things, we can treat it all as data and depend on the
application-specific schema to namespace block data vs state data.

Note that we still end up with hash, parentHash, number, and weight
as block metadata that Storage requires to function, but any
additional chain-specific block metadata can be stored in the same
way as state data with an appropriate namespace.