Add ZeroCopyGet
Some database engines support zero copy reads, in which it is safe
to interact with the data returned by the database, but not okay
to modify it. That memory may be reused in subsequent requests, making
it unsafe to keep a reference past a certain point.

This introduces a ZeroCopyGet function at both the storage.db and
storage.current interface in which the value is provided to a closure
in which it is safe to manipulate the data, but beyond which
references to the data should not be retained.

This aligns nicely with how BadgerDB implements zero copy reads, but
is probably a good way to approach it if we do zero copy reads in
LMDB or other database engines.