Prevent some nasty race conditions
Previously, if AddBlock() was called during the life of a View,
there was a risk that the view could return incorrect information.

The view held a database transaction that would always represent a
particular snapshot of the database. But when AddBlock() was called,
the subsequent consolidate() would change some memoryLayer.parent
from another memory layer to the disk layer. This meant that the
original memoryLayer.parent would get skipped by reads, but the
database transaction wouldn't reflect the merged in data, thus any
data that changed in that layer wouldn't be reflected after AddBlock().

This makes a shallow copy of the memoryLayer() stack for each view,
which won't be updated during the consolidate() step, offering
consistency.

The locks used here aren't held for any extended periods. The views
use RLocks only for the period of time that the stack of memory
layers is being copied and the database transaction is created. The
consolidate process holds the write lock only for long enough to
update the reference to l.parent.

Additionally, I added a lock used by AddBlock() to ensure that only
one AddBlock() can be running at a time. Current cardinal
implementations will have only a single thread calling AddBlock(),
so this shouldn't be necessary, but I don't know what crazy things
people might try to do in the future.