Add caching on isDeleted lookups
isDeleted does some simple string manipulation (splitting and joining)
but because it has to be done at every layer for every lookup, this
adds up to a fair bit of computation time. Using a cache for the
splitting / joining means that compute only needs to be done once,
instead of at every layer, which is a substantial performance improvement.