Strip memory layers that don't connect to the disk layer on resumption
When Cardinal terminates without persisting memory layers to disk, the
memory layers may be disconnected from the current disk layer. If we
get an `AddBlock()` call on top of a layer that is disconnected from
the disk layer, the consolidation process panics and crashes.

This change strips out any layers that don't connect to the disk layer.
If Cardinal processed > 128 blocks before shutting down, this will likely
be all of them, in which case resumption will need to proceed from the
disk layer without the persisted memory layers, but that should be fine.

It also prevents the disk layer from being overwritten in the map in the
event that the disk layer's hash is present in the persisted memory layers.
This would occur if Cardinal processed < 128 blocks before shutting down.