Update s.latestHash inside lock
We ran into a race condition where a request could get s.latestHash
after it was updated in AddBlock() but before the new layer was added
to s.layers. This could result in ErrLayerNotFound errors for users
requesting "latest" as the block, which should never occur.

Moving the update of s.latestHash inside the lock context ensures
that, while other threads may get s.latestHash before the new layer
is in s.layers (because s.LatestHash() is not locked), any calls
to view will wait until the new layer is in s.layers before attempting
the lookup.