Use WriteBatches for potentially large writes
In BadgerDB (and possibly others in the future) there is a limit to
the amount that can be written in a single atomic write. Previously,
we would have simply lost any changes beyond the threshold.

This change adds a parallel for BadgerDB's WriteBatch, which can
automatically break up writes that exceed the limit into multiple
writes, to insure all writes are completed. Note that if writes
exceed the size of a single transaction, we will lose atomicity,
but that's generally better than silently failing to write further updates.