Storage Resolver
Now that we support multiple persistence backends (badger and bolt) as
well as multiple storage engines (current and archive), we need to
be able to resolve database engines.

This implementation simply points to a path on disk, resolves what kind
of persistence backend exists at that location, and what kind of storage
engine it uses, then returns a storage.Storage object of the appropriate
type.