Update rollback to use resolver
The rollback subcommand now uses the resolver. This revealed that
the resolver had some compile-time issues, which have now been addressed.

Additionally, I came to realize that because code no longer needs a
reference to the underlying database, it's harder to close. Thus,
the storage engine will now close the underlying database.