Handle deletes before handling puts
In Ethereum, "destructs" and contract creations can happen at the
same time, particularly in cases where an address was funded before
a contract was deployed at that location.

Doing updates then deletes results in the final state being that
an address does not exist, while the final state should be a
deployed contract.

This also updates the delta / rollback implementation to account
for the possiblity that both deletes and updates happen at the same
time.