Memory layers: Check deletes after current level
We just adjusted disk writes to do deletes first then writes so
that updated values will be reflected instead of deletions. We need
the same behavior when checking memory layers - if something is
deleted and updated in the same block, it should return the updated
value instead of showing it as deleted.