archiveLayer.getIndex fix
archiveLayer.getIndex didn't have correct logic for finding an index
for a given block. It's been a long time since I first implemented
that logic, and I'm not entirely sure what I was thinking, but
ultimately Rank() should be the index we want to use unless it's 0,
in which case we can consider the block not found because that means
the first entry in the index is greater than the requested block.