Implement waiters
This allows applications to register Cardinal Streams waiters with
Cardinal Storage, ensuring that if requests want to run against blocks
that are in the middle of processing that the Cardinal Storage engine
will wait for the blocks to finish processing before returning results.

This implements in both the Current and Archive storage. Applications have
to call s.RegisterWaiter(watier, timeout) for waiters to take effect - a
null waiter will be in place otherwise.

Both current and archive tests have been updated to account for waiters,
and additional tests confirm waiter functionality is working as intended.